refactored friendlist
